rename _ports_utils.py to ports_utils.py and add docstring
